+++ src/python-common/ceph/deployment/service_spec.py
@@ -31,6 +31,7 @@ from typing import (
     TypedDict,
     Literal
 )
+from venv import logger
 
 import yaml
